+++ src/libcamera/sensor/meson.build
@@ -2,5 +2,6 @@
 
 libcamera_internal_sources += files([
     'camera_sensor.cpp',
+    'camera_sensor_legacy.cpp',
     'camera_sensor_properties.cpp',
 ])
